Homework 02 - работа над оибками
