+++ homework02/src/main/java/ru/otus/spring02/Main.java
@@ -1,29 +1,11 @@
 package ru.otus.spring02;
 
-import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.*;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import ru.otus.spring02.service.ExamService;
 
 @PropertySource("classpath:application.properties")
-@Configuration
 @ComponentScan
 public class Main {
-
-    @Bean
-    public static PropertySourcesPlaceholderConfigurer propertyConfig(){
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource source = new ReloadableResourceBundleMessageSource();
-        source.setBasename("/i18n/bundle");
-        source.setDefaultEncoding("UTF-8");
-        return source;
-    }
-
     public static void main(String[] args) {
         AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
         ExamService exam = context.getBean(ExamService.class);
