+++ homework02/src/main/java/ru/otus/spring02/config/AppConfig.java
@@ -1,0 +1,23 @@
+package ru.otus.spring02.config;
+
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+@Configuration
+public class AppConfig {
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer propertyConfig(){
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+    @Bean
+    public MessageSource messageSource() {
+        ReloadableResourceBundleMessageSource source = new ReloadableResourceBundleMessageSource();
+        source.setBasename("/i18n/bundle");
+        source.setDefaultEncoding("UTF-8");
+        return source;
+    }
+}
